Implement internal validator: counts solution costs and violations

+++ src/config.h
@@ -1,9 +1,9 @@
 #ifndef CONFIG_H
 #define CONFIG_H
 
-static const double ROOM_CAPACITY_COST = 1;
-static const double MIN_WORKING_DAYS_COST = 5;
-static const double CURRICULUM_COMPACTNESS_COST = 2;
-static const double ROOM_STABILITY_COST = 1;
+static const int ROOM_CAPACITY_COST = 1;
+static const int MIN_WORKING_DAYS_COST = 5;
+static const int CURRICULUM_COMPACTNESS_COST = 2;
+static const int ROOM_STABILITY_COST = 1;
 
 #endif // CONFIG_H
